requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
google-generativeai>=0.8.0
google-api-core>=2.0.0
PyMuPDF>=1.24.0
requests>=2.31.0

# Server Dependencies (optional - for server.py)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
